Refactor and Structuring Code
Folderisasi sesuai fungsi, hapus endpoint gadipake
