+++ CC-Path/app/config.py
@@ -1,0 +1,10 @@
+DATABASE_CONFIG = {
+    "user": "root",
+    "password": "C23-PS076",
+    "host": "34.101.247.215",
+    "database": "moger"
+}
+
+SECRET_KEY = "your-secret-key"
+ALGORITHM = "HS256"
+ACCESS_TOKEN_EXPIRE_MINUTES = 30
