+++ CC-Path/app/dependencies/database.py
@@ -1,0 +1,9 @@
+from app.database.connection import get_database_connection, connection_pool
+
+# Dependency to get a database connection
+def get_database_connection():
+    connection = connection_pool.get_connection()
+    try:
+        yield connection
+    finally:
+        connection.close()
