+++ CC-Path/app/models/user.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+
+class User(BaseModel):
+    email: str
+    name: str
+    password: str
+
+class UserLogin(BaseModel):
+    email: str
+    password: str
