+++ CC-Path/app/utils/password.py
@@ -1,0 +1,11 @@
+from passlib.context import CryptContext
+
+pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+
+# Generate hashed password
+def get_password_hash(password):
+    return pwd_context.hash(password)
+
+# Verify password
+def verify_password(plain_password, hashed_password):
+    return pwd_context.verify(plain_password, hashed_password)
